+++ ginger_webhook.php
@@ -1,6 +1,4 @@
-<?php$method_name = strtoupper(explode('.php',basename(__FILE__))[0]);
-
-
+<?php
 require_once('includes/application_top.php');
 require_once('includes/classes/gateways/autoload.php');